[#389] Move JUnitXMLReport
* Expose the building blocks and utilities.
